enhance Transistion: add resizable observer
